requirements.txt
bugsnag~=4.2.1
mysql-connector-python~=8.0.29
aiogram~=2.19
asyncio~=3.4.3
aioflask~=0.4.0
asgiref~=3.5.2